requirements.txt
requests
PyYAML
numpy